move pvica test methods
