+++ src/test/java/dk/kb/present/transform/XSLTTransformerTestBase.java
@@ -27,8 +27,6 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.function.Consumer;
 
-import static dk.kb.present.solr.EmbeddedSolrTest.PRESERVICA2SOLR;
-import static dk.kb.present.transform.XSLTPreservicaSchemaOrgTransformerTest.PRESERVICA2SCHEMAORG;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -66,18 +64,6 @@ public abstract class XSLTTransformerTestBase {
     }
 
     /**
-     * Wrapper for {@link #assertMultiTestsThroughSchemaTransformation(String, Consumer[])} which verifies that the
-     * transformed record contains the given {@code substring}.
-     * @param recordFile the file to load, transform and test.
-     * @param substring must be present in the transformed record.
-     */
-    public void assertPvicaContains(String recordFile, String substring) {
-        assertMultiTestsThroughSchemaTransformation(recordFile,
-                solrDoc -> assertTrue(solrDoc.contains(substring))
-        );
-    }
-
-    /**
      * Wrapper for {@link #assertMultiTests(String, Consumer[])} that verifies that the transformed record contains
      * the given {@code substring}.
      * @param recordFile the file to load, transform and test.
@@ -99,18 +85,6 @@ public abstract class XSLTTransformerTestBase {
     public void assertNotContains(String recordFile, String substring) {
         // TODO: When mods2solr.xslt has been converted into schemaorg2solr.xslt combine the assertNotContains methods
         assertMultiTests(recordFile,
-                solrDoc -> assertFalse(solrDoc.contains(substring))
-        );
-    }
-
-    /**
-     * Wrapper for {@link #assertMultiTestsThroughSchemaTransformation(String, Consumer[])} which verifies that the
-     * transformed record does not contain the given {@code substring}.
-     * @param recordFile the file to load, transform and test.
-     * @param substring must be present in the transformed record.
-     */
-    public void assertPvicaNotContains(String recordFile, String substring) {
-        assertMultiTestsThroughSchemaTransformation(recordFile,
                 solrDoc -> assertFalse(solrDoc.contains(substring))
         );
     }
@@ -161,30 +135,4 @@ public abstract class XSLTTransformerTestBase {
         Arrays.stream(tests).forEach(test -> test.accept(solrString));
     }
 
-    /**
-     * Checks that internal test files are available and if not, logs a warning and returns.
-     * <p>
-     * If the check passes, the content of the file {@code record} is transformed using two XSLTs.
-     * At first the XML record is transformed to Schema.org JSON and then the schema.org JSON is transformed to solr
-     * documents and the given tests are performed on the result.
-     * @param record file with a record that is to be transformed.
-     * @param tests Zero or more tests to perform on the transformed record.
-     */
-    @SafeVarargs
-    public final void assertMultiTestsThroughSchemaTransformation(String record, Consumer<String>... tests){
-        if (!TestFileProvider.hasSomeTestFiles()) {
-            return;  // ensureTestFiles takes care of logging is there are no internal test files
-        }
-        String solrString;
-        try {
-            solrString = TestUtil.getTransformedToSolrJsonThroughSchemaJson(PRESERVICA2SCHEMAORG, record);
-        } catch (Exception e) {
-            throw new RuntimeException(
-                    "Unable to fetch and transform '" + record + "' using XSLT '" + getXSLT() + "'", e);
-        }
-
-        //TestUtil.prettyPrintJson(solrString);
-
-        Arrays.stream(tests).forEach(test -> test.accept(solrString));
-    }
 }
